requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt
python-jose[cryptography]
python-multipart